refactor: 변수명, 변수타입 변경

+++ planus_back/src/main/java/com/planus/websocket/model/WebSocketMessage.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class WebSocketMessage {
-    private double tripId;
+    private long tripId;
     private String userName;
     private String chatMsg;
-    private int action; // 0:입장, ...
+    private int action; // 0:입장, 1:채팅 ...
 
 }
